# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: UNILIV Admin API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /employees:
    get:
      operationId: getEmployees
      tags: [employees]
      summary: List employees
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
        - $ref: "#/components/parameters/SearchParam"
        - $ref: "#/components/parameters/PropertyIdParam"
        - in: query
          name: department
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
      responses:
        "200":
          description: Employees list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeesListResponse"
    post:
      operationId: createEmployee
      tags: [employees]
      summary: Create employee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEmployeeBody"
      responses:
        "201":
          description: Employee created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeeResponse"
  /employees/{id}:
    get:
      operationId: getEmployee
      tags: [employees]
      summary: Get employee by ID
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Employee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeeResponse"
    put:
      operationId: updateEmployee
      tags: [employees]
      summary: Update employee
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEmployeeBody"
      responses:
        "200":
          description: Employee updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeeResponse"
    delete:
      operationId: deleteEmployee
      tags: [employees]
      summary: Delete employee
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ─── ATTENDANCE ───────────────────────────────────────
...
  /job-requisitions:
    get:
      operationId: getJobRequisitions
      tags: [hrRecruitment]
      summary: List job requisitions
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
      responses:
        "200":
          description: Job requisitions list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobRequisitionsListResponse"
    post:
      operationId: createJobRequisition
      tags: [hrRecruitment]
      summary: Create job requisition
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateJobRequisitionBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobRequisitionResponse"
  /candidates:
    get:
      operationId: getCandidates
      tags: [hrRecruitment]
      summary: List candidates
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
        - $ref: "#/components/parameters/SearchParam"
        - in: query
          name: stage
          schema:
            type: string
      responses:
        "200":
          description: Candidates list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CandidatesListResponse"
    post:
      operationId: createCandidate
      tags: [hrRecruitment]
      summary: Create candidate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCandidateBody"
      responses:
        "201":
          description: Candidate created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CandidateResponse"
  /candidates/{id}:
    put:
      operationId: updateCandidate
      tags: [hrRecruitment]
      summary: Update candidate
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCandidateBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CandidateResponse"

  # ─── VENDORS ──────────────────────────────────────────
...
components:
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 1
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
    SearchParam:
      name: search
      in: query
      schema:
        type: string
    PropertyIdParam:
      name: propertyId
      in: query
      schema:
        type: string
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required: [total, page, limit, totalPages]

    # ─── AUTH ─────────────────────────────────────────────
...
    EmployeeDto:
      type: object
      properties:
        id:
          type: string
        employeeCode:
          type: string
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        dob:
          type: string
          nullable: true
        gender:
          type: string
          nullable: true
        photo:
          type: string
          nullable: true
        department:
          type: string
        designation:
          type: string
        propertyId:
          type: string
          nullable: true
        managerId:
          type: string
          nullable: true
        joiningDate:
          type: string
        ctc:
          type: number
          nullable: true
        status:
          type: string
        createdAt:
          type: string
      required: [id, employeeCode, name, email, phone, department, designation, joiningDate, status, createdAt]
    EmployeeResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/EmployeeDto"
      required: [success, data]
    EmployeesListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: "#/components/schemas/EmployeeDto"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
      required: [success, data, meta]
    CreateEmployeeBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        dob:
          type: string
        gender:
          type: string
        department:
          type: string
        designation:
          type: string
        propertyId:
          type: string
        managerId:
          type: string
        joiningDate:
          type: string
        ctc:
          type: number
        bankAccount:
          type: string
        ifscCode:
          type: string
        panNumber:
          type: string
        status:
          type: string
      required: [name, email, phone, department, designation, joiningDate]

    # ─── ATTENDANCE ───────────────────────────────────────
...
    JobRequisitionDto:
      type: object
      properties:
        id:
          type: string
        role:
          type: string
        department:
          type: string
        headcount:
          type: integer
        status:
          type: string
        candidateCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, role, department, headcount, status, candidateCount, createdAt, updatedAt]
    JobRequisitionResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/JobRequisitionDto"
      required: [success, data]
    JobRequisitionsListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: "#/components/schemas/JobRequisitionDto"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
      required: [success, data, meta]
    CreateJobRequisitionBody:
      type: object
      properties:
        role:
          type: string
        department:
          type: string
        headcount:
          type: integer
        status:
          type: string
      required: [role, department, headcount]
    CandidateDto:
      type: object
      properties:
        id:
          type: string
        jobRequisitionId:
          type: string
          nullable: true
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        resumeUrl:
          type: string
          nullable: true
        source:
          type: string
          nullable: true
        stage:
          type: string
        bgvStatus:
          type: string
          nullable: true
        offerStatus:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, name, email, phone, stage, createdAt]
    CandidateResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/CandidateDto"
      required: [success, data]
    CandidatesListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: "#/components/schemas/CandidateDto"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
      required: [success, data, meta]
    CreateCandidateBody:
      type: object
      properties:
        jobRequisitionId:
          type: string
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        resumeUrl:
          type: string
        source:
          type: string
        stage:
          type: string
        notes:
          type: string
      required: [name, email, phone]

    # ─── VENDOR ───────────────────────────────────────────
...