# lib/api-zod/src/generated/types/createCandidateBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * UNILIV Admin API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateCandidateBody {
  jobRequisitionId?: string;
  name: string;
  email: string;
  phone: string;
  resumeUrl?: string;
  source?: string;
  stage?: string;
  notes?: string;
}